Address owner review: dead kick code, conftest location, phase select_related
- member/views.py: remove transfer_management_if_needed call from
  MemberKickView — the manager can never be kicked (only the manager
  can kick, and self-kicks are blocked), so this was dead code
- conftest.py: move four management-transfer factories from the per-app
  game/tests/conftest.py to the root service/conftest.py per CLAUDE.md
  convention; delete the now-empty per-app conftest
- phase/models.py: add game__managing_member to with_adjudication_data
  select_related so that game.managing_member.user_id is served from
  cache when transfer_management_if_needed is called during resolution
