+++ service/member/views.py
@@ -56,11 +56,9 @@ class MemberKickView(SelectedGameMixin, generics.DestroyAPIView):
 
     def perform_destroy(self, instance):
         game = instance.game
-        kicked_user = instance.user
         user_id = instance.user_id
         with transaction.atomic():
             instance.delete()
-            game.transfer_management_if_needed(kicked_user, reason="was kicked")
             if user_id:
                 send_notification.defer(
                     user_ids=[user_id],
